+++ packages/oruga-next/src/components/collapse/Collapse.vue
@@ -1,7 +1,7 @@
 <script lang="ts">
 import { defineComponent, h, Transition, vShow, withDirectives } from "vue";
 
-import BaseComponentMixin from "../../utils/BaseComponentMixin";
+import BaseComponentMixin from "../../mixins/BaseComponentMixin";
 import { getOptions } from "../../utils/config";
 import { getValueByPath } from "../../utils/helpers";
