+++ packages/oruga-next/src/components/datepicker/DatepickerTableRow.vue
@@ -1,7 +1,7 @@
 <script lang="ts">
 import { defineComponent, type PropType } from "vue";
 
-import BaseComponentMixin from "../../utils/BaseComponentMixin";
+import BaseComponentMixin from "../../mixins/BaseComponentMixin";
 
 export default defineComponent({
     name: "ODatepickerTableRow",
