+++ packages/oruga-next/src/components/skeleton/Skeleton.vue
@@ -1,7 +1,7 @@
 <script lang="ts">
 import { defineComponent, h } from "vue";
 
-import BaseComponentMixin from "../../utils/BaseComponentMixin";
+import BaseComponentMixin from "../../mixins/BaseComponentMixin";
 import { toCssDimension } from "../../utils/helpers";
 
 /**
