+++ packages/oruga-next/src/mixins/MatchMediaMixin.ts
@@ -1,6 +1,6 @@
 import { defineComponent } from "vue";
-import { getOptions } from "./config";
-import { getValueByPath } from "./helpers";
+import { getOptions } from "../utils/config";
+import { getValueByPath } from "../utils/helpers";
 
 export default defineComponent({
     props: {
